Working hours views

+++ app/src/main/java/com/aliakseipilko/flightdutytracker/view/fragment/hoursFragments/base/BaseHourFragment.java
@@ -4,9 +4,9 @@ import android.support.v4.app.Fragment;
 
 public abstract class BaseHourFragment extends Fragment {
 
-    public abstract void showDutyHours(String s, long dutyMillis);
+    public abstract void showDutyHours(String s, double dutyMillis);
 
-    public abstract void showFlightHours(String s, long flightMillis);
+    public abstract void showFlightHours(String s, double flightMillis);
 
     public abstract void showError(String message);
 }
